Remove server.abc module
