--- server/abc/__init__.py
@@ -1,3 +1,0 @@
-"""
-Abstract base classes
-"""
