+++ server/gameconnection.py
@@ -10,12 +10,19 @@ from sqlalchemy import select, text
 
 from server.db import FAFDatabase
 
-from .abc.base_game import GameConnectionState
 from .config import TRACE
 from .db.models import coop_leaderboard, coop_map, teamkills
 from .decorators import with_logger
 from .game_service import GameService
-from .games import CoopGame, Game, GameError, GameState, ValidityState, Victory
+from .games import (
+    CoopGame,
+    Game,
+    GameConnectionState,
+    GameError,
+    GameState,
+    ValidityState,
+    Victory
+)
 from .games.typedefs import FA
 from .player_service import PlayerService
 from .players import Player, PlayerState
