+++ server/lobbyconnection.py
@@ -20,7 +20,6 @@ from sqlalchemy import and_, func, select
 import server.metrics as metrics
 from server.db import FAFDatabase
 
-from .abc.base_game import GameConnectionState, InitMode
 from .config import TRACE, config
 from .db.models import (
     avatars,
@@ -40,7 +39,9 @@ from .games import (
     CoopGame,
     CustomGame,
     FeaturedModType,
+    GameConnectionState,
     GameState,
+    InitMode,
     VisibilityState
 )
 from .geoip_service import GeoIpService
