+++ tests/unit_tests/test_lobbyconnection.py
@@ -9,13 +9,12 @@ from aiohttp import web
 from asynctest import CoroutineMock
 from sqlalchemy import and_, select
 
-from server.abc.base_game import InitMode
 from server.config import config
 from server.db.models import ban, friends_and_foes
 from server.exceptions import BanError, ClientError
 from server.game_service import GameService
 from server.gameconnection import GameConnection
-from server.games import CustomGame, Game, GameState, VisibilityState
+from server.games import CustomGame, Game, GameState, InitMode, VisibilityState
 from server.geoip_service import GeoIpService
 from server.ice_servers.nts import TwilioNTS
 from server.ladder_service import LadderService
